Add api django app * add vacancies django app * do refactor * delete config.py and add pothon-dotenv
