+++ .gitignore
@@ -131,4 +131,4 @@ dmypy.json
 # Some stuff
 .idea
 .vscode
-config.py
+django_site/django_site/config.py
