+++ django_site/parser_vacancies/management/commands/config_parser.py
@@ -3,14 +3,14 @@ REQUEST_PER_PAGE = 100
 # Количество запрашиваемых страниц
 REQUEST_PAGE_COUNT = 20
 # Базовый URL для запроса вакансий у api hh.ru
-REQUEST_URL = f'https://api.hh.ru/vacancies?text=python&per_page={REQUEST_PER_PAGE}&page='
+REQUEST_URL = f'https://api.hh.ru/vacancies?text=python&area=1&per_page={REQUEST_PER_PAGE}&page='
 # Список обрабатываемых скиллов
 PROCESSING_SKILLS = [
     'sql', 'linux', 'git', 'postgresql', 'django framework', 'django', 'flask',
     'css', 'html', 'bash', 'mysql', 'ооп', 'docker', 'mongodb', 'pandas',
     'numpy', 'atlassian jira', 'jira', 'kubernetes', 'http', 'tcp/ip',
     'trello', 'unix', 'redis'
-                    ]
+]
 # Минимальная длинна скилла
 min_lenght_skill_letter = 0
 for letter in PROCESSING_SKILLS:
