+++ django_site/recommendation/config_recommendation.py
@@ -1,5 +1,6 @@
-DEFAULT_SKILL_NAME = ['sql', 'linux', 'git', 'postgresql', 'django', 'flask', 'css', 'html', 'bash', 'mysql', 'oop', 'docker',
-                      'mongodb', 'pandas', 'numpy', 'jira', 'kubernetes', 'http', 'tcp_ip', 'trello', 'unix', 'redis'
-                     ]
+DEFAULT_SKILL_NAME = [
+    'sql', 'linux', 'git', 'postgresql', 'django', 'flask', 'css', 'html', 'bash', 'mysql', 'oop',
+    'docker', 'mongodb', 'pandas', 'numpy', 'jira', 'kubernetes', 'http', 'tcp_ip', 'trello', 'unix', 'redis'
+]
 
 NUM_TOP_VACANCIES = 5
