+++ django_site/users/models.py
@@ -1,15 +1,15 @@
-from django.db import models
-from django.core.validators import MaxValueValidator, MinValueValidator
 from django.contrib.auth.models import AbstractBaseUser
 from django.contrib.auth.models import PermissionsMixin
-from django.utils.translation import gettext_lazy as _
+from django.db import models
 from django.utils import timezone
+from django.utils.translation import gettext_lazy as _
 
 from .managers import UserManager
 
 
 class User(AbstractBaseUser, PermissionsMixin):
-    ''' Таблица для юзеров.'''
+    """ Таблица для юзеров."""
+
     additional_skills = 'Через запятую укажите дополнительные технологии, '
     additional_skills += 'с которыми у вас есть опыт работы.'
     text = 'Многие работодатели не указывают уровень зп, показывать такие вакансии?'
@@ -47,37 +47,4 @@ class User(AbstractBaseUser, PermissionsMixin):
         return self.email
 
 
-class Vacancies(models.Model):
-    ''' Таблица в бд для всех вакансий с парсинга.'''
-    shown_to_users = models.ManyToManyField(User, related_name='shown_vacancies')
-    banned_by_users = models.ManyToManyField(User, related_name='banned_vacancies')
-    id_vacancy = models.CharField(max_length=100, unique=True, blank=True)
-    name = models.CharField(max_length=100, blank=True)
-    area = models.CharField(max_length=100, blank=True)
-    experience = models.CharField(max_length=100, blank=True)
-    salary_from = models.CharField(max_length=100, blank=True)
-    salary_to = models.CharField(max_length=100, blank=True)
-    url = models.CharField(max_length=100, blank=True)
-    published = models.DateTimeField(blank=True)
-    contains_skills = models.BooleanField(blank=True, null=True)
 
-    class Meta:
-        app_label = 'users'
-        managed = True
-
-    def __repr__(self):
-        return self.id_vacancy
-
-
-class Rating(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE, blank=True, null=True)
-    vacancy = models.ForeignKey(Vacancies, on_delete=models.CASCADE, blank=True, null=True)
-    rating = models.IntegerField(default=0,
-                                 validators=[
-                                     MaxValueValidator(5),
-                                     MinValueValidator(0),
-                                 ]
-                                 )
-
-    def __str__(self):
-        return str(self.pk)
