+++ django_site/users/views.py
@@ -1,19 +1,20 @@
-from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.views.decorators.http import require_http_methods
-from django.views.decorators.csrf import csrf_exempt
-from django.http import HttpResponse
 from django.core.paginator import Paginator
-from django.http import JsonResponse
+from django.http import HttpResponse
+from django.shortcuts import render, redirect
+from django.views.decorators.csrf import csrf_exempt
+from django.views.decorators.http import require_http_methods
+from vacancies.models import Vacancies, Rating
 
 from .forms import UserCreationForm, UserChangeForm
-from .models import User, Rating, Vacancies
+from .models import User
 from .services import profile_view, remove_user_from_vacancy_relation
 
 
 def join(request):
-    ''' Регистрации юзера с последующим редиректом на страницу логина.'''
+    """ Регистрации юзера с последующим редиректом на страницу логина."""
+
     if request.method == 'POST':
         form = UserCreationForm(request.POST)
         if form.is_valid():
@@ -32,7 +33,8 @@ def join(request):
 
 @login_required
 def profile(request):
-    ''' Профиль юзера с выводом вакансий для него.'''
+    """ Профиль юзера с выводом вакансий для него."""
+
     user_request = User.objects.filter(email=request.user).values('id',
                                                                   'area',
                                                                   'salary',
@@ -57,7 +59,8 @@ def profile(request):
 
 @login_required
 def edit_profile(request):
-    ''' Изменение данных профиля в лк.'''
+    """ Изменение данных профиля в лк."""
+
     if request.method == 'POST':
         edit_form = UserChangeForm(request.POST, instance=request.user)
         if edit_form.is_valid():
@@ -81,7 +84,8 @@ def edit_profile(request):
 @require_http_methods(["POST"])
 @csrf_exempt
 def rate_vacancy(request):
-    ''' Оценка вакансии пользователем '''
+    """ Оценка вакансии пользователем """
+
     rating = request.POST.get('rate')
     vacancy = Vacancies.objects.get(pk=request.POST.get('vacancy'))
     user = User.objects.get(pk=request.user.id)
