+++ django_site/vacancies/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class VacanciesConfig(AppConfig):
+    name = 'vacancies'
