+++ requirements.txt
@@ -12,3 +12,4 @@ requests==2.25.0
 soupsieve==2.1
 sqlparse==0.4.1
 urllib3==1.26.2
+python-dotenv==0.15.0
